requirements.txt
alchemy-mock==0.4.3
atomicwrites==1.4.0
attrs==21.4.0
bcrypt==3.2.0
cffi==1.15.0
click==8.0.4
colorama==0.4.3
coverage==6.3.2
cryptography==37.0.4
flake8==4.0.1
Flask==2.0.3
flask-accept==0.0.6
Flask-Bcrypt==0.7.1
Flask-Login==0.4.1
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.14.2
greenlet==1.1.2
iniconfig==1.1.1
itsdangerous==2.1.0
Jinja2==3.0.3
MarkupSafe==2.1.0
marshmallow==3.14.1
mccabe==0.6.1
packaging==21.3
pluggy==1.0.0
py==1.11.0
pycodestyle==2.8.0
pycparser==2.18
pyflakes==2.4.0
PyJWT==2.4.0
pyparsing==3.0.7
pytest==7.0.1
pytest-cov==3.0.0
six==1.11.0
SQLAlchemy==1.4.31
tomli==2.0.1
Werkzeug==2.0.3
WTForms==2.1
